+++ src/components/Author/AuthorDetail.js
@@ -1,4 +1,4 @@
-import { API } from "../api.js";
+import { API } from "../../api.js";
 import React from 'react';
 import { withRouter } from "react-router-dom";
